requirements.txt
# RecognitionSystem - Modern Dependencies
# Updated for 2025/2026 compatibility

# Core ML Frameworks
tensorflow>=2.15.0
torch>=2.1.0
torchaudio>=2.1.0

# Transformers & NLP
transformers>=4.36.0
sentencepiece>=0.1.99
tokenizers>=0.15.0

# Speech Emotion Recognition (Modern)
# speechbrain>=1.0.0  # Optional: Alternative to custom models

# Video/Face Analysis
deepface>=0.0.93
mediapipe>=0.10.9
opencv-python>=4.9.0.80

# Video Download (Replaces deprecated youtube-dl)
yt-dlp>=2024.1.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
audioread>=3.0.1

# PDF Processing
PyPDF2>=3.0.1

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# PyQt5 (Desktop UI - Legacy support)
PyQt5>=5.15.10

# Video Processing
moviepy>=1.0.3
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Web Framework (API Layer)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async Processing
celery>=5.3.0
redis>=5.0.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
Pillow>=10.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0

